chore: Update .p10k.zsh configuration
The .p10k.zsh configuration file has been updated. This change reflects the latest configuration generated by the Powerlevel10k configuration wizard on 2023-10-23 at 14:30 EDT. The update is based on the romkatv/powerlevel10k/config/p10k-lean.zsh template with checksum 45810.

Changes Made:
- Updated the timestamp to the current date and time.

This commit ensures that the configuration is up-to-date and accurately represents the current state of the Powerlevel10k setup.