feat: Update Neovim configuration
This commit includes changes to the Neovim configuration file and related Lua scripts. The updates are as follows:

- Added the 'VeryLazy' event to the checker configuration for plugin updates.
- Set 'lazy' properties for several plugins, including 'nvim-lspconfig,' 'nvim-cmp,' 'nvim-autopairs,' and 'nvim-dap,' to improve performance.
- Adjusted the configuration for 'nvim-dap' and 'nvim-dap-ui' to enhance the debugging experience.
- Tweaked the TokyoNight theme settings for Neovim.

These changes are intended to improve the overall performance and functionality of the Neovim setup. The 'VeryLazy' event is now used for plugin updates, and various plugins have been configured for lazy loading to optimize startup times.