chore: Lazy-load and optimize Neovim plugins
This commit optimizes the plugin loading process in the Neovim configuration by enabling lazy loading for specific plugins. It also ensures that certain plugins are loaded only when working with Lua files for a more efficient workflow. The changes include:

- Lazy-loading of plugins:
  - 'mg979/vim-visual-multi' will now be loaded with the 'VeryLazy' event.
  - 'tpope/vim-sleuth' is configured for lazy loading.
  - 'nvim-tree/nvim-web-devicons' is set to lazy load.
  - 'williamboman/mason.nvim' is configured for lazy loading with its dependencies.
  - 'j-hui/fidget.nvim' will be loaded with the 'LspAttach' event.
  - 'folke/neodev.nvim' is now configured to load only for Lua files.
  - 'hrsh7th/nvim-cmp' is set to load on 'InsertEnter' event.
  - 'folke/which-key.nvim' is now loaded lazily with 'VeryLazy'.
  - 'numToStr/Comment.nvim' is configured to load with the 'VeryLazy' event.
  - 'nvim-telescope/telescope.nvim' is optimized for lazy loading with its dependencies.
  - 'cbochs/portal.nvim' is configured to load lazily with 'VeryLazy'.

These changes enhance the startup performance of Neovim and ensure that plugins are loaded only when needed, improving the overall efficiency of the Neovim environment.