chore: Update Neovim plugin configurations
This commit updates the configuration of Neovim plugins to include the 'event = 'VeryLazy'' option. This change is made to enhance performance and make the plugins more efficient, especially when Neovim is in a very low activity state.

- Added 'event = 'VeryLazy'' to the configurations of multiple plugins.

These changes optimize the behavior of the plugins and contribute to a smoother user experience in Neovim.