refactor: autoformat.lua
In this commit, the autoformat.lua script has been refactored to improve its clarity and readability. Specifically, the following changes were made:

- Line 47: Commented out the block of code related to the 'tsserver' client, as it was causing issues with bad language support. This block of code is now enclosed in a multiline comment.

These changes aim to make the autoformat.lua script more maintainable and allow for better handling of the 'tsserver' client. The script is now more straightforward for those working with it.

I may uncomment this setting depending if the server starts working poorly