style: Adjust Neovim and Wezterm colorschemes
In the Neovim configuration, updated the MatchParen highlighting to use standout with a customized background and foreground color. Changed the CursorLineNr highlighting to have a bold style and specified the foreground color while keeping the background transparent. Tweaked the TelescopeSelection, Visual, and LineNr highlights for better visual elements.

In the Wezterm configuration, added a cursor_bg color setting with the value '#E5E9F0' to adjust the cursor background color.