chore: exclude PHP from custom keymaps and Treesitter plugins
In the core keymaps configuration, PHP is included in the list of excluded filetypes to prevent specific keymaps from being applied when working with PHP files.

In the Treesitter plugin configuration, PHP is added to the list of ignored installations to avoid attempting to install Treesitter support for PHP files.