refactor(install_tool): update ohmyzsh installation
This commit updates the install_tool script to automate the installation
of Oh My Zsh and keeps the existing zshrc configuration. The
`--keep-zshrc` option is added to the Oh My Zsh installation command,
preventing the replacement of the current zshrc file. The validation for
zsh config is adjusted accordingly. Additionally, the `chsh` command to
set zsh as the default shell is removed to maintain manual control over
the shell configuration.