refactor(nvim): nvim on wegovy
This commit does the following but not limited to:
- created autocommands to load plugins conditionally
- removed stylistic plugins like nvim-web-devicons for less bloat
- removed after/ftplugins for less bloat
- moved the initializing into the the lua/iton folder to have better
  access
- skimmed down the lsp and completion to just have the bare essentials
- renamed globals.lua -> utils.lua to for clarity
- removed unused plugins like debug.lua (may reinstall later but
  currently no use case)
- changed lazy.nvim UI to be very basic

This commit tries to make neovim as minimal and modular with the bare
essentials that I currently require. Additionally, it allows for better
cherry-picking between main and wsl branch since it no longer uses
Nerd Fonts. I want to use NeoVim as a new Vim rather than a bare bone VS Code.